Remove `Key`
