+++ src/moscot/base/problems/problem.py
@@ -20,12 +20,12 @@ from sklearn.preprocessing import StandardScaler
 
 import scanpy as sc
 
-from moscot._constants._constants import ProblemStage
 from moscot._docs._docs import d
 from moscot._types import ArrayLike, CostFn_t, Device_t
 from moscot.base.output import BaseSolverOutput
 from moscot.base.problems._utils import require_solution, wrap_prepare, wrap_solve
 from moscot.base.solver import OTSolver, ProblemKind
+from moscot.constants import ProblemStage
 from moscot.logging import logger
 from moscot.utils._tagged_array import Tag, TaggedArray
