+++ src/moscot/problems/space/_mixins.py
@@ -27,12 +27,12 @@ from scipy.stats import pearsonr, spearmanr
 from sklearn.metrics import pairwise_distances
 from sklearn.neighbors import NearestNeighbors
 
-from moscot._constants._constants import AlignmentMode, CorrMethod, PlottingDefaults
 from moscot._docs._docs import d
 from moscot._docs._docs_mixins import d_mixins
 from moscot._types import ArrayLike, Device_t, Str_Dict_t
 from moscot.base.problems._mixins import AnalysisMixin, AnalysisMixinProtocol
 from moscot.base.problems.compound_problem import B, K
+from moscot.constants import AlignmentMode, CorrMethod, PlottingDefaults
 from moscot.logging import logger
 from moscot.utils._subset_policy import StarPolicy
